Support Node.js ESM.
Fixes https://github.com/zeit/styled-jsx/issues/640 .
