+++ css.mjs
@@ -1,0 +1,5 @@
+import decoys from './dist/decoys.js'
+
+export default decoys.css
+export const global = decoys.global
+export const resolve = decoys.resolve
