+++ test/babel6/snapshots/attribute.js.md
@@ -8,7 +8,7 @@ Generated by [AVA](https://ava.li).
 
 > Snapshot 1
 
-    `import _JSXStyle from 'styled-jsx/style';␊
+    `import _JSXStyle from 'styled-jsx/style.js';␊
     import styles from './styles';␊
     ␊
     const styles2 = require('./styles2');␊
@@ -108,7 +108,7 @@ Generated by [AVA](https://ava.li).
 
     `var _this = this;␊
     ␊
-    import _JSXStyle from "styled-jsx/style";␊
+    import _JSXStyle from "styled-jsx/style.js";␊
     export default (() => {␊
       const Element = 'div';␊
       return <div className={"jsx-2886504620"}>␊
