+++ test/babel6/snapshots/external.js.md
@@ -8,7 +8,7 @@ Generated by [AVA](https://ava.li).
 
 > Snapshot 1
 
-    `import _JSXStyle from 'styled-jsx/style';␊
+    `import _JSXStyle from 'styled-jsx/style.js';␊
     ␊
     import colors, { size } from './constants';␊
     const color = 'red';␊
@@ -98,7 +98,7 @@ Generated by [AVA](https://ava.li).
 
 > Snapshot 1
 
-    `import _JSXStyle from 'styled-jsx/style';␊
+    `import _JSXStyle from 'styled-jsx/style.js';␊
     ␊
     ␊
     function test() {␊
@@ -112,7 +112,7 @@ Generated by [AVA](https://ava.li).
 
 > Snapshot 1
 
-    `import _JSXStyle from 'styled-jsx/style';␊
+    `import _JSXStyle from 'styled-jsx/style.js';␊
     ␊
     import colors, { size } from './constants';␊
     const color = 'red';␊
@@ -168,7 +168,7 @@ Generated by [AVA](https://ava.li).
 
 > Snapshot 1
 
-    `import _JSXStyle from "styled-jsx/style";␊
+    `import _JSXStyle from "styled-jsx/style.js";␊
     import styles from './styles2';␊
     ␊
     export default (({ level = 1 }) => {␊
@@ -184,7 +184,7 @@ Generated by [AVA](https://ava.li).
 
 > Snapshot 1
 
-    `import _JSXStyle from 'styled-jsx/style';␊
+    `import _JSXStyle from 'styled-jsx/style.js';␊
     import styles from './styles';␊
     const styles2 = require('./styles2');␊
     import { foo as styles3 } from './styles';␊
@@ -211,7 +211,7 @@ Generated by [AVA](https://ava.li).
 
 > Snapshot 1
 
-    `import _JSXStyle from 'styled-jsx/style';␊
+    `import _JSXStyle from 'styled-jsx/style.js';␊
     import styles, { foo as styles3 } from './styles';␊
     ␊
     const styles2 = require('./styles2');␊
@@ -229,7 +229,7 @@ Generated by [AVA](https://ava.li).
 
 > Snapshot 1
 
-    `import _JSXStyle from 'styled-jsx/style';␊
+    `import _JSXStyle from 'styled-jsx/style.js';␊
     import styles from './styles';␊
     ␊
     export default (() => <div className={`jsx-${styles.__hash}`}>␊
