+++ test/babel6/snapshots/index.js.md
@@ -8,7 +8,7 @@ Generated by [AVA](https://ava.li).
 
 > Snapshot 1
 
-    `import _JSXStyle from 'styled-jsx/style';␊
+    `import _JSXStyle from 'styled-jsx/style.js';␊
     export default (() => <div className={"jsx-2743241663"}>␊
         <p className={"jsx-2743241663"}>test</p>␊
         <p className={"jsx-2743241663"}>woot</p>␊
@@ -28,7 +28,7 @@ Generated by [AVA](https://ava.li).
 
 > Snapshot 1
 
-    `import _JSXStyle from 'styled-jsx/style';␊
+    `import _JSXStyle from 'styled-jsx/style.js';␊
     export default (() => <div className={"jsx-2743241663"}>␊
         <p className={"jsx-2743241663"}>test</p>␊
         <p className={"jsx-2743241663"}>woot</p>␊
@@ -40,7 +40,7 @@ Generated by [AVA](https://ava.li).
 
 > Snapshot 1
 
-    `import _JSXStyle from 'styled-jsx/style';␊
+    `import _JSXStyle from 'styled-jsx/style.js';␊
     const Test = () => <_JSXStyle id={"2743241663"}>{"p{color:red;}"}</_JSXStyle>;␊
     ␊
     export default (() => <div className={"jsx-2673076688"}>␊
@@ -53,7 +53,7 @@ Generated by [AVA](https://ava.li).
 
 > Snapshot 1
 
-    `import _JSXStyle from 'styled-jsx/style';␊
+    `import _JSXStyle from 'styled-jsx/style.js';␊
     const color = 'red';␊
     const otherColor = 'green';␊
     ␊
@@ -76,7 +76,7 @@ Generated by [AVA](https://ava.li).
 
 > Snapshot 1
 
-    `import _JSXStyle from "styled-jsx/style";␊
+    `import _JSXStyle from "styled-jsx/style.js";␊
     const Test = () => <div className={"jsx-2529315885"}>␊
         <span className={"jsx-2529315885"}>test</span>␊
         <Component />␊
@@ -87,7 +87,7 @@ Generated by [AVA](https://ava.li).
 
 > Snapshot 1
 
-    `import _JSXStyle from "styled-jsx/style";␊
+    `import _JSXStyle from "styled-jsx/style.js";␊
     export default class {␊
     ␊
       render() {␊
@@ -103,7 +103,7 @@ Generated by [AVA](https://ava.li).
 
 > Snapshot 1
 
-    `import _JSXStyle from 'styled-jsx/style';␊
+    `import _JSXStyle from 'styled-jsx/style.js';␊
     ␊
     ␊
     export default (({ children }) => <div className={`jsx-${styles.__hash}`}>␊
@@ -127,7 +127,7 @@ Generated by [AVA](https://ava.li).
 
 > Snapshot 1
 
-    `import _JSXStyle from "styled-jsx/style";␊
+    `import _JSXStyle from "styled-jsx/style.js";␊
     export default (({ level = 1 }) => {␊
       const Element = `h${level}`;␊
     ␊
@@ -158,7 +158,7 @@ Generated by [AVA](https://ava.li).
 
 > Snapshot 1
 
-    `import _JSXStyle from 'styled-jsx/style';␊
+    `import _JSXStyle from 'styled-jsx/style.js';␊
     export default class {␊
       render() {␊
         const Element = 'div';␊
@@ -184,7 +184,7 @@ Generated by [AVA](https://ava.li).
 
 > Snapshot 1
 
-    `import _JSXStyle from 'styled-jsx/style';␊
+    `import _JSXStyle from 'styled-jsx/style.js';␊
     const darken = c => c;␊
     const color = 'red';␊
     const otherColor = 'green';␊
@@ -215,7 +215,7 @@ Generated by [AVA](https://ava.li).
 
 > Snapshot 1
 
-    `import _JSXStyle from 'styled-jsx/style';␊
+    `import _JSXStyle from 'styled-jsx/style.js';␊
     const Test = () => <div className={"jsx-2209073070"}>␊
         <_JSXStyle id={"2209073070"}>{"body{color:red;}:hover{color:red;display:-webkit-box;display:-webkit-flex;display:-ms-flexbox;display:flex;-webkit-animation:foo 1s ease-out;animation:foo 1s ease-out;}div a{display:none;}[data-test]>div{color:red;}"}</_JSXStyle>␊
       </div>;␊
@@ -226,7 +226,7 @@ Generated by [AVA](https://ava.li).
 
 > Snapshot 1
 
-    `import _JSXStyle from "styled-jsx/style";␊
+    `import _JSXStyle from "styled-jsx/style.js";␊
     const attrs = {␊
       id: 'test'␊
     };␊
@@ -257,7 +257,7 @@ Generated by [AVA](https://ava.li).
 
 > Snapshot 1
 
-    `import _JSXStyle from 'styled-jsx/style';␊
+    `import _JSXStyle from 'styled-jsx/style.js';␊
     export default (() => <div className={"jsx-2743241663"}>␊
         <p className={"jsx-2743241663"}>woot</p>␊
         <style dangerouslySetInnerHTML={{ __html: `body { margin: 0; }` }}></style>␊
@@ -268,7 +268,7 @@ Generated by [AVA](https://ava.li).
 
 > Snapshot 1
 
-    `import _JSXStyle from 'styled-jsx/style';␊
+    `import _JSXStyle from 'styled-jsx/style.js';␊
     export default (() => <div className={"jsx-2743241663"}>␊
         <p className={"jsx-2743241663"}>test</p>␊
         <p className={"jsx-2743241663"}>woot</p>␊
