+++ test/babel6/snapshots/macro.js.md
@@ -8,7 +8,7 @@ Generated by [AVA](https://ava.li).
 
 > Snapshot 1
 
-    `import _JSXStyle from 'styled-jsx/style';␊
+    `import _JSXStyle from 'styled-jsx/style.js';␊
     ␊
     ␊
     ({␊
@@ -20,7 +20,7 @@ Generated by [AVA](https://ava.li).
 
 > Snapshot 1
 
-    `import _JSXStyle from 'styled-jsx/style';␊
+    `import _JSXStyle from 'styled-jsx/style.js';␊
     ␊
     ␊
     function test() {␊
@@ -34,7 +34,7 @@ Generated by [AVA](https://ava.li).
 
 > Snapshot 1
 
-    `import _JSXStyle from 'styled-jsx/style';␊
+    `import _JSXStyle from 'styled-jsx/style.js';␊
     ␊
     ␊
     const { className, styles } = {␊
