+++ test/babel6/snapshots/plugins.js.md
@@ -8,7 +8,7 @@ Generated by [AVA](https://ava.li).
 
 > Snapshot 1
 
-    `import _JSXStyle from 'styled-jsx/style';␊
+    `import _JSXStyle from 'styled-jsx/style.js';␊
     import styles from './styles';␊
     const color = 'red';␊
     ␊
@@ -39,7 +39,7 @@ Generated by [AVA](https://ava.li).
 
 > Snapshot 1
 
-    `import _JSXStyle from 'styled-jsx/style';␊
+    `import _JSXStyle from 'styled-jsx/style.js';␊
     import styles from './styles';␊
     const color = 'red';␊
     ␊
