+++ test/snapshots/external.js.md
@@ -8,7 +8,7 @@ Generated by [AVA](https://ava.li).
 
 > Snapshot 1
 
-    `import _JSXStyle from "styled-jsx/style";␊
+    `import _JSXStyle from "styled-jsx/style.js";␊
     import colors, { size } from './constants';␊
     const color = 'red';␊
     const bar = ["div.jsx-2141779268{font-size:3em;}"];␊
@@ -55,7 +55,7 @@ Generated by [AVA](https://ava.li).
 
     `"use strict";␊
     ␊
-    var _style = _interopRequireDefault(require("styled-jsx/style"));␊
+    var _style = _interopRequireDefault(require("styled-jsx/style.js"));␊
     ␊
     var _App = _interopRequireDefault(require("./App.styles"));␊
     ␊
@@ -99,7 +99,7 @@ Generated by [AVA](https://ava.li).
 
 > Snapshot 1
 
-    `import _JSXStyle from "styled-jsx/style";␊
+    `import _JSXStyle from "styled-jsx/style.js";␊
     ␊
     function test() {␊
       ({␊
@@ -112,7 +112,7 @@ Generated by [AVA](https://ava.li).
 
 > Snapshot 1
 
-    `import _JSXStyle from "styled-jsx/style";␊
+    `import _JSXStyle from "styled-jsx/style.js";␊
     import colors, { size } from './constants';␊
     const color = 'red';␊
     const bar = new String("div.jsx-2141779268{font-size:3em;}");␊
@@ -158,7 +158,7 @@ Generated by [AVA](https://ava.li).
 
 > Snapshot 1
 
-    `import _JSXStyle from "styled-jsx/style";␊
+    `import _JSXStyle from "styled-jsx/style.js";␊
     import styles from './styles2';␊
     export default (({␊
       level = 1␊
@@ -174,7 +174,7 @@ Generated by [AVA](https://ava.li).
 
 > Snapshot 1
 
-    `import _JSXStyle from "styled-jsx/style";␊
+    `import _JSXStyle from "styled-jsx/style.js";␊
     import styles from './styles';␊
     ␊
     const styles2 = require('./styles2');␊
@@ -201,7 +201,7 @@ Generated by [AVA](https://ava.li).
 
 > Snapshot 1
 
-    `import _JSXStyle from "styled-jsx/style";␊
+    `import _JSXStyle from "styled-jsx/style.js";␊
     import styles, { foo as styles3 } from './styles';␊
     ␊
     const styles2 = require('./styles2');␊
@@ -219,7 +219,7 @@ Generated by [AVA](https://ava.li).
 
 > Snapshot 1
 
-    `import _JSXStyle from "styled-jsx/style";␊
+    `import _JSXStyle from "styled-jsx/style.js";␊
     import styles from './styles';␊
     export default (() => <div className={`jsx-${styles.__hash}`}>␊
         <p className={`jsx-${styles.__hash}`}>test</p>␊
