+++ test/snapshots/index.js.md
@@ -4,11 +4,41 @@ The actual snapshot is saved in `index.js.snap`.
 
 Generated by [AVA](https://ava.li).
 
+## Babel plugin matches import specifier styled-jsx/css
+
+> Snapshot 1
+
+    `import _JSXStyle from "styled-jsx/style.js";␊
+    ({␊
+      styles: <_JSXStyle id={"2886504620"}>{"div.jsx-2886504620{color:red;}"}</_JSXStyle>,␊
+      className: "jsx-2886504620"␊
+    });`
+
+## Babel plugin matches import specifier styled-jsx/css.js
+
+> Snapshot 1
+
+    `import _JSXStyle from "styled-jsx/style.js";␊
+    ({␊
+      styles: <_JSXStyle id={"2886504620"}>{"div.jsx-2886504620{color:red;}"}</_JSXStyle>,␊
+      className: "jsx-2886504620"␊
+    });`
+
+## Babel plugin matches import specifier styled-jsx/css.mjs
+
+> Snapshot 1
+
+    `import _JSXStyle from "styled-jsx/style.js";␊
+    ({␊
+      styles: <_JSXStyle id={"2886504620"}>{"div.jsx-2886504620{color:red;}"}</_JSXStyle>,␊
+      className: "jsx-2886504620"␊
+    });`
+
 ## generates source maps
 
 > Snapshot 1
 
-    `import _JSXStyle from "styled-jsx/style";␊
+    `import _JSXStyle from "styled-jsx/style.js";␊
     export default (() => <div className={"jsx-2743241663"}>␊
         <p className={"jsx-2743241663"}>test</p>␊
         <p className={"jsx-2743241663"}>woot</p>␊
@@ -19,7 +49,7 @@ Generated by [AVA](https://ava.li).
 
 > Snapshot 1
 
-    `import _JSXStyle from "styled-jsx/style";␊
+    `import _JSXStyle from "styled-jsx/style.js";␊
     import React, { Component } from 'react';␊
     export default class Index extends Component {␊
       static getInitialProps() {␊
@@ -72,7 +102,7 @@ Generated by [AVA](https://ava.li).
 
 > Snapshot 1
 
-    `import _JSXStyle from "styled-jsx/style";␊
+    `import _JSXStyle from "styled-jsx/style.js";␊
     export default (() => <div className={"jsx-2743241663"}>␊
         <p className={"jsx-2743241663"}>test</p>␊
         <p className={"jsx-2743241663"}>woot</p>␊
@@ -84,7 +114,7 @@ Generated by [AVA](https://ava.li).
 
 > Snapshot 1
 
-    `import _JSXStyle from "styled-jsx/style";␊
+    `import _JSXStyle from "styled-jsx/style.js";␊
     ␊
     const Test = () => <_JSXStyle id={"2743241663"}>{"p{color:red;}"}</_JSXStyle>;␊
     ␊
@@ -98,7 +128,7 @@ Generated by [AVA](https://ava.li).
 
 > Snapshot 1
 
-    `import _JSXStyle from "styled-jsx/style";␊
+    `import _JSXStyle from "styled-jsx/style.js";␊
     const color = 'red';␊
     const otherColor = 'green';␊
     ␊
@@ -121,7 +151,7 @@ Generated by [AVA](https://ava.li).
 
 > Snapshot 1
 
-    `import _JSXStyle from "styled-jsx/style";␊
+    `import _JSXStyle from "styled-jsx/style.js";␊
     ␊
     const Test = () => <div className={"jsx-2529315885"}>␊
         <span className={"jsx-2529315885"}>test</span>␊
@@ -133,7 +163,7 @@ Generated by [AVA](https://ava.li).
 
 > Snapshot 1
 
-    `import _JSXStyle from "styled-jsx/style";␊
+    `import _JSXStyle from "styled-jsx/style.js";␊
     export default class {␊
       render() {␊
         return <div className={"jsx-2101845350"}>␊
@@ -148,7 +178,7 @@ Generated by [AVA](https://ava.li).
 
 > Snapshot 1
 
-    `import _JSXStyle from "styled-jsx/style";␊
+    `import _JSXStyle from "styled-jsx/style.js";␊
     export default (({␊
       children␊
     }) => <div className={`jsx-${styles.__hash}`}>␊
@@ -172,7 +202,7 @@ Generated by [AVA](https://ava.li).
 
 > Snapshot 1
 
-    `import _JSXStyle from "styled-jsx/style";␊
+    `import _JSXStyle from "styled-jsx/style.js";␊
     export default (({␊
       level = 1␊
     }) => {␊
@@ -203,7 +233,7 @@ Generated by [AVA](https://ava.li).
 
 > Snapshot 1
 
-    `import _JSXStyle from "styled-jsx/style";␊
+    `import _JSXStyle from "styled-jsx/style.js";␊
     export default class {␊
       render() {␊
         const Element = 'div';␊
@@ -229,7 +259,7 @@ Generated by [AVA](https://ava.li).
 
 > Snapshot 1
 
-    `import _JSXStyle from "styled-jsx/style";␊
+    `import _JSXStyle from "styled-jsx/style.js";␊
     ␊
     const darken = c => c;␊
     ␊
@@ -265,7 +295,7 @@ Generated by [AVA](https://ava.li).
 
 > Snapshot 1
 
-    `import _JSXStyle from "styled-jsx/style";␊
+    `import _JSXStyle from "styled-jsx/style.js";␊
     import React from 'react';␊
     export default (() => <>␊
         <p className={"jsx-1712488456"}>Testing!!!</p>␊
@@ -300,7 +330,7 @@ Generated by [AVA](https://ava.li).
 
 > Snapshot 1
 
-    `import _JSXStyle from "styled-jsx/style";␊
+    `import _JSXStyle from "styled-jsx/style.js";␊
     ␊
     const Test = () => <div className={"jsx-2209073070"}>␊
         <_JSXStyle id={"2209073070"}>{"body{color:red;}:hover{color:red;display:-webkit-box;display:-webkit-flex;display:-ms-flexbox;display:flex;-webkit-animation:foo 1s ease-out;animation:foo 1s ease-out;}div a{display:none;}[data-test]>div{color:red;}"}</_JSXStyle>␊
@@ -312,7 +342,7 @@ Generated by [AVA](https://ava.li).
 
 > Snapshot 1
 
-    `import _JSXStyle from "styled-jsx/style";␊
+    `import _JSXStyle from "styled-jsx/style.js";␊
     const attrs = {␊
       id: 'test'␊
     };␊
@@ -344,7 +374,7 @@ Generated by [AVA](https://ava.li).
 
 > Snapshot 1
 
-    `import _JSXStyle from "styled-jsx/style";␊
+    `import _JSXStyle from "styled-jsx/style.js";␊
     export default (() => <div className={"jsx-2743241663"}>␊
         <p className={"jsx-2743241663"}>woot</p>␊
         <style dangerouslySetInnerHTML={{␊
@@ -357,10 +387,17 @@ Generated by [AVA](https://ava.li).
 
 > Snapshot 1
 
-    `import _JSXStyle from "styled-jsx/style";␊
+    `import _JSXStyle from "styled-jsx/style.js";␊
     export default (() => <div className={"jsx-2743241663"}>␊
         <p className={"jsx-2743241663"}>test</p>␊
         <p className={"jsx-2743241663"}>woot</p>␊
         <p className={"jsx-2743241663"}>woot</p>␊
         <_JSXStyle id={"2743241663"}>{"p.jsx-2743241663{color:red;}"}</_JSXStyle>␊
       </div>);`
+
+## Babel plugin doesn’t match an unrelated import specifier
+
+> Snapshot 1
+
+    `import css from 'x';␊
+    css.resolve`div { color: red }`;`
