+++ test/snapshots/plugins.js.md
@@ -8,7 +8,7 @@ Generated by [AVA](https://ava.li).
 
 > Snapshot 1
 
-    `import _JSXStyle from "styled-jsx/style";␊
+    `import _JSXStyle from "styled-jsx/style.js";␊
     import styles from './styles';␊
     const color = 'red';␊
     export default (() => <div className={"jsx-3382438999 " + `jsx-${styles.__hash}`}>␊
@@ -37,7 +37,7 @@ Generated by [AVA](https://ava.li).
 
 > Snapshot 1
 
-    `import _JSXStyle from "styled-jsx/style";␊
+    `import _JSXStyle from "styled-jsx/style.js";␊
     import styles from './styles';␊
     const color = 'red';␊
     export default (() => <div className={"jsx-3382438999 " + `jsx-${styles.__hash}`}>␊
